Handling Errors

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -30,14 +30,11 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            //policy
-            $this->isAble('store', User::class);;
+        if ($this->isAble('store', User::class)) {
 
             return new UserResource(User::create($request->mappedAttributes()));
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to create that resource', 401);
         }
+        return $this->notAuthorized('You are not authorized to create that resource');
     }
 
     /**
@@ -55,57 +52,41 @@ class UserController extends ApiController
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateUserRequest $request, $user_id)
+    public function update(UpdateUserRequest $request, User $user)
     {
-        try {
-            $user = User::findOrFail($user_id);
-
-            //policy
-            $this->isAble('update', $user);
+        // PATCH
+        if ($this->isAble('update', $user)) {
 
             $user->update($request->mappedAttributes());
 
             return new UserResource($user);
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
+        return $this->notAuthorized('You are not authorized to update that resource');
     }
 
-    public function replace(ReplaceUserRequest $request, $user_id)
+    public function replace(ReplaceUserRequest $request, User $user)
     {
         // PUT
-        try {
-            $user = User::findOrFail($user_id);
-
-            //policy
-            $this->isAble('replace', $user);
+        if ($this->isAble('replace', $user)) {
 
             $user->update($request->mappedAttributes());
 
             return new UserResource($user);
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found', 404);
         }
+        return $this->notAuthorized('You are not authorized to replace that resource');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($user_id)
+    public function destroy(User $user)
     {
-        try {
-            $user = User::findOrFail($user_id);
-
-            //policy
-            $this->isAble('delete', $user);
+        if($this->isAble('delete', $user)) {
 
             $user->delete();
 
             return $this->ok('User successfully deleted');
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found', 404);
         }
+        return $this->notAuthorized('You are not authorized to delete that resource');
     }
 }
